Move form field class in the right folder
